Rename `note_expression.rs` to `note_expressions.rs` for consistency
