+++ plinth-plugin/src/formats/vst3.rs
@@ -5,7 +5,7 @@ mod factory;
 mod host;
 mod key_codes;
 mod macros;
-mod note_expression;
+mod note_expressions;
 mod parameters;
 mod plugin;
 mod stream;
